requirements.txt
asgiref==3.4.1
Django==3.2.20
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-jsonapi==4.3.0
inflection==0.5.1
mysqlclient==1.4.2.post1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
unicodecsv==0.14.1
